requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
python-dotenv==1.0.1
Werkzeug==3.0.1 